telegram-bot/requirements.txt
# Telegram Bot
telethon==1.33.0

# WhatsApp Bot (Walichat)
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0

# Shared dependencies
anthropic>=0.40.0
supabase>=2.10.0
python-dotenv==1.0.0
PyPDF2>=3.0.0
gspread>=5.12.0
google-auth>=2.23.0
